pyproject.toml
[project]
name = "fastapi-buzz"
version = "0.1.0"
description = "Extra bindings on py-buzz specifically for fastapi apps"
authors = [
    {name = "Tucker Beck", email ="tucker.beck@gmail.com"},
]
readme = "README.md"
license-files = ["LICENSE.md"]
requires-python = ">=3.10"
dependencies = [
    "py-buzz~=7.0",
    "fastapi>=0.100.0,<1.0",
]

[project.urls]
homepage = "https://github.com/dusktreader/fastapi-buzz"
source = "https://github.com/dusktreader/fastapi-buzz"
changelog = "https://github.com/dusktreader/fastapi-buzz/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "basedpyright~=1.28",
    "fastapi[standard]~=0.115",
    "httpx>=0.28.1",
    "ipython~=8.35",
    "mkdocs-material~=9.6",
    "mkdocstrings[python]~=0.29",
    "mypy~=1.15",
    "pyclean~=3.1",
    "pygments~=2.19",
    "pytest~=8.3",
    "pytest-cov~=6.1",
    "pytest-random-order~=1.1",
    "ruff~=0.11",
    "typos~=1.31",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
addopts = [
    "--random-order",
    "--cov=src/fastapi_buzz",
    "--cov-report=term-missing",
    "--cov-fail-under=90",
    "--cov-report=xml:.coverage.xml",
    "--junitxml=.junit.xml",
    "--override-ini=junit_family=legacy",
]

[tool.ruff]
line-length = 120
src = ["src/fastapi_buzz", "tests"]

[tool.mypy]
pretty = true
files = ["src/fastapi_buzz", "tests"]

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportUnusedCallResult = false

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_buzz"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
